+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-## [0.8.0] - in development
+## [0.8.0] - 2023-12-07
 
 ### [Changelog][0.8.0-changes]
 
 ### Added
 
-- Jump forward/backward by number of lines, or by time interval in seconds, minutes, hours or days.
+- Jump forward/backward by number of lines, or by time interval in microseconds, milliseconds, seconds, minutes, hours or days.
 - Support for `--inline` command-line option, to merge logs into a single column instead of side by side.
 
 ### Fixed
@@ -130,7 +130,7 @@
 [0.2.0]: https://github.com/ptmcg/log_merger/releases/tag/v0.2.0
 [0.1.0]: https://github.com/ptmcg/log_merger/releases/tag/v0.1.0
 
-[0.8.0-changes]: https://github.com/ptmcg/log_merger/compare/v0.7.0...main
+[0.8.0-changes]: https://github.com/ptmcg/log_merger/compare/v0.7.0...v0.8.0
 [0.7.0-changes]: https://github.com/ptmcg/log_merger/compare/v0.6.0...v0.7.0
 [0.6.0-changes]: https://github.com/ptmcg/log_merger/compare/v0.5.0...v0.6.0
 [0.5.0-changes]: https://github.com/ptmcg/log_merger/compare/v0.4.0...v0.5.0